Factor out component loading into its own crate
