+++ Makefile.toml
@@ -13,6 +13,7 @@ CARGO_MAKE_CRATE_WORKSPACE_MEMBERS = [
 	"iocuddle",
 	"iocuddle-sgx",
 	"keep-runtime",
+	"loader",
 	"memory",
 	"payload",
 	"sallyport",
