+++ loader/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "loader"
+version = "0.1.0"
+authors = ["Connor Kuehl <ckuehl@redhat.com>"]
+edition = "2018"
+license = "Apache-2.0"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+goblin = "0.2.3"
+libc = "0.2.71"
+mmap = { path = "../mmap" }
+memory = { path = "../memory" }
+span = { path = "../span" }
